balancing + bug fixes
alpaca have ~1 base dps, those with aoe have slightly less.
modifier cards generally provide 1.5-2x damage increase(or some equivalent value in utility ie slow).
with max stack size of 5, each alpaca have max damage output of around 10-15/sec ideally, though some combinations work better than others.
made modifier pack more expensive than alpaca pack since they offer more dps value.

alpaca no longer attacks while being dragged.
aoe from projectile no longer hit bugs already hit by the initial projectile.

+++ Assets/Tyler/Scripts/Modifiers/AOEModifier.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class AOEModifier : MonoBehaviour, ICardModifier
 {
-    [SerializeField] int value = 1;
+    [SerializeField] float value = 1;
 
     public void OnActivate(Alpaca alpaca)
     {
